+++ CHANGELOG.md
@@ -25,6 +25,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.0] - 2022-10-05
+
+### Changed
+- Client code for accessing CSM has been factored out into the `csm-api-client`
+  library. `cfs-config-util` now uses `csm-api-client` as a dependency for
+  accessing CSM APIs. The Python modules
+  `cfs_config_util.{cfs,vcs,session,apiclient}` have been removed and replaced
+  with the corresponding modules in the `csm_api_client.service` package. The
+  `cfs_config_util.bin.vcs_creds_helper` module has been removed.
+
 ## [3.3.1] - 2022-09-30
 
 ### Security